Phase 6: split control panels into focused components, typed visitLoop
- components/program/: ProgramControlPanel container + ProgramToolbar,
  CodeEditor, CodeInsertMenu, BlocklyEditor (mockable boundary)
- components/world/: WorldControlPanel container + WorldToolbar
- grammar: label the three loop alternatives (WhileLoop/DoWhileLoop/
  DoUntilLoop); visitLoop now dispatches via instanceof on the generated
  contexts and iterates the typed statement() lists instead of scanning
  child texts for 'solange'/'bis' (also stops re-evaluating the loop
  condition as part of the body children)
